Add files via upload
Well its cost our sanity... but a continued update to the GUI calculator. Please note that to run with CSV right now you must access the package contents and run the python file directly. Why? I'm not really sure. I've tried everything including, and not limited to, summoning ancient beasts of the land, but nothing seems to stick. Hopefully we can sort this out before the age of governments comes to an end and the AI overlords holds us in virtual prisons of our minds. Enjoy!